SIL: Stored property initializers get the linkage of their containing type
This allows inlineble initializers to be defined on types that
have stored property initializers.

This is not the final behavior we want; in Swift 4 mode, we want
to give these shared linkage and make them fragile, and enforce
that stored property initializer expressions only reference other
public and versioned symbols.

This will match the behavior of default argument expressions in
Swift 4 mode, and I'll implement it when I do those.