utils/update-checkout: Do any ``git checkout`` before the ``git fetch``. utils/update-checkout: Add a call to ``git rebase --abort`` when using the ``--clean`` option. Calling ``git fetch`` updates .git/FETCH_HEAD and marks all branches besides the currently checked out branch as not-for-merge. So tot ensure that the branch we want to merge is available for merging, we must do a checkout of that branch before updating the FETCH_HEAD file.
We also must ensure that we have fetched before optionally resetting to remote.

Fixes https://bugs.swift.org/browse/SR-3800.

The --clean option is meant to restore your repository to a pristine
condition, but there are cases where you can ``reset --hard`` and still be
in the middle of a rebase. This is annoying across 10+ repositories, so
if the user wants to --clean, then abort all rebases in progress.